refactor(server): share the credential-less GraphQL refusal between the guards
UserAuthGuard and WorkspaceAuthGuard carried the same getRequest / no-request /
throw-UNAUTHENTICATED preamble verbatim, down to the AuthenticationError
arguments, so a change to how a credential-less request is reported had to be
made in both. That is the drift this stack exists to close, reintroduced one
level up.

getRequestOrThrowWhenUnauthenticated now owns it and both guards call it, each
keeping only what it actually decides: a user for one, a workspace for the
other. hasAuthenticatedPrincipal folds into that util as a local predicate,
since the guards were its only consumers.
