--- packages/twenty-server/src/engine/utils/has-authenticated-principal.util.ts
@@ -1,14 +1,0 @@
-import { isDefined } from 'twenty-shared/utils';
-
-export type AuthenticatablePrincipalRequest = {
-  user?: unknown;
-  apiKey?: unknown;
-  application?: unknown;
-};
-
-export const hasAuthenticatedPrincipal = (
-  request: AuthenticatablePrincipalRequest,
-): boolean =>
-  isDefined(request.user) ||
-  isDefined(request.apiKey) ||
-  isDefined(request.application);
